==> chap_19/06_PWV.ipynb <==
# Sentiment Analysis with Pretrained Word Vectors

### Loading Libraries

# Numerical Computing
import numpy as np

# Data Manipulation
import pandas as pd
import pandas_datareader.data as web

# Data Visualization
import seaborn as sns
import matplotlib.pyplot as plt
from matplotlib.ticker import FuncFormatter

# Warning
import warnings

# Path
from pathlib import Path

# Scikit-Learn
from sklearn.metrics import roc_auc_score

# TensorFlow
import tensorflow as tf
import tensorflow.keras.backend as K
from tensorflow.keras.models import Sequential
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.layers import Dense, GRU, Embedding
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.callbacks import ModelCheckpoint, EarlyStopping

np.random.seed(42)

idx = pd.IndexSlice

sns.set_style('whitegrid')

gpu_devices = tf.config.experimental.list_physical_devices('GPU')

if gpu_devices:
    print('Using GPU')
    tf.config.experimental.set_memory_growth(gpu_devices[0], True)
else:
    print('Using CPU')

results_path = Path('results', 'sentiment_imdb')

if not results_path.exists():
    results_path.mkdir(parents=True)

### Loading Reviews

path = Path('aclImdb')

files = path.rglob('*.txt')

data = []

for f in files:
    if f.stem.startswith(('urls_', 'imdbEr')):
        continue

    parts = f.parent.parts

    if len(parts) != 3:
        continue

    _, data_set, outcome = parts
    if outcome == 'unsup':
        continue

    review = f.read_text(encoding='latin1')
    data.append([data_set, int(outcome == 'pos'), review])

data = pd.DataFrame(data, columns=['dataset', 'label', 'review'])
data.info()

train_data = data.loc[data.dataset=='train', ['label', 'review']]

test_data = data.loc[data.dataset=='test', ['label', 'review']]

train_data.label.value_counts()

test_data.label.value_counts()

### Preparing Data

#### Tokenizer

num_words = 10000

t = Tokenizer(num_words=num_words, 
              lower=True, 
              oov_token=2)

t.fit_on_texts(train_data.review)

vocab_size = len(t.word_index) + 1

vocab_size

train_data_encoded = t.texts_to_sequences(train_data.review)

test_data_encoded = t.texts_to_sequences(test_data.review)

max_length = 100

### Padding Sequences

# Trained Padded Sequences
X_train_padded = pad_sequences(train_data_encoded, 
                            maxlen=max_length, 
                            padding='post',
                           truncating='post')

y_train = train_data['label']

X_train_padded.shape

# Tested Padded Sequences
X_test_padded = pad_sequences(test_data_encoded, 
                            maxlen=max_length, 
                            padding='post',
                           truncating='post')

y_test = test_data['label']

X_test_padded.shape

### Loading Embeddings

glove_path = Path('glove', 'glove.6B.100d.txt')

embeddings_index = dict()

for line in glove_path.open(encoding='latin1'):
    values = line.split()
    word = values[0]
    try:
        coefs = np.asarray(values[1:], dtype='float32')
    except:
        continue
    embeddings_index[word] = coefs

print('Loaded {:,d} word vectors.'.format(len(embeddings_index)))

embedding_matrix = np.zeros((vocab_size, 100))

for word, i in t.word_index.items():
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector

embedding_matrix.shape

### Defining Model Architecture

embedding_size = 100

# Building Network
rnn = Sequential([
    Embedding(input_dim=vocab_size, 
              output_dim= embedding_size, 
              input_length=max_length,
              weights=[embedding_matrix], 
              trainable=False),
    GRU(units=32,  dropout=0.2, recurrent_dropout=0.2),
    Dense(1, activation='sigmoid')
])

rnn.build(input_shape=(None, max_length))
rnn.summary()

# Compiling Network
rnn.compile(loss='binary_crossentropy',
            optimizer='RMSProp',
            metrics=['accuracy', 
                     tf.keras.metrics.AUC(name='AUC')])

rnn_path = (results_path / 'lstm.pretrained.h5').as_posix()

checkpointer = ModelCheckpoint(filepath=rnn_path,
                               verbose=1,
                               monitor='val_AUC',
                               mode='max',
                               save_best_only=True)

early_stopping = EarlyStopping(monitor='val_AUC',
                               patience=5,
                               mode='max',
                               restore_best_weights=True)

training = rnn.fit(X_train_padded,
                   y_train,
                   batch_size=32,
                   epochs=100,
                   validation_data=(X_test_padded,
                                    y_test),
                   callbacks=[early_stopping,
                              checkpointer],
                   verbose=1)

# Getting Prediction Score
y_score = rnn.predict(X_test_padded)

roc_auc_score(y_score=y_score.squeeze(), y_true=y_test)

df = pd.DataFrame(training.history)

best_auc = df.val_AUC.max()
best_acc = df.val_accuracy.max()

fig, axes = plt.subplots(ncols=2, figsize=(14,4))
df.index = df.index.to_series().add(1)
df[['AUC', 'val_AUC']].plot(ax=axes[0], 
                            title=f'AUC | Best: {best_auc:.4f}', 
                            legend=False, 
                            xlim=(1, 33),
                            ylim=(.7, .95))

axes[0].axvline(df.val_AUC.idxmax(), ls='--', lw=1, c='k')
df[['accuracy', 'val_accuracy']].plot(ax=axes[1], 
                                              title=f'Accuracy | Best: {best_acc:.2%}', 
                                              legend=False, 
                                              xlim=(1, 33),
                                      ylim=(.7, .9))
axes[1].axvline(df.val_accuracy.idxmax(), ls='--', lw=1, c='k')
axes[0].set_xlabel('Epoch')
axes[0].set_ylabel('AUC')
axes[1].set_xlabel('Epoch')
axes[1].set_ylabel('Accuracy')
fig.suptitle('Sentiment Analysis - Pretrained Vectors', fontsize=14)
fig.legend(['Train', 'Validation'], loc='center right')

sns.despine()
fig.tight_layout()
fig.subplots_adjust(top=.9)
fig.savefig(results_path / 'imdb_pretrained', dpi=300);
plt.grid()
plt.show()
